Refactor inventory management functions

+++ Shigger/inventory.lua
@@ -33,12 +33,17 @@ function inventory.checkInventory()
         if config.empty_thrash then
             local thrash_exists = false
             for i=2, 16 do
-                for _, thrash in ipairs(thrash_list) do
-                    if turtle.getItemDetail(i) then
-                        if turtle.getItemDetail(i).name:find(thrash, 1 ,true) then
-                            turtle.select(i)
-                            turtle.drop()
-                            thrash_exists = true
+                if turtle.getItemDetail(i) then
+                    for _, whitelist in ipairs(config.whitelist) do
+                        if turtle.getItemDetail(i).name:find(whitelist, 1 ,true) then
+                            break
+                        end
+                        for _, thrash in ipairs(thrash_list) do
+                            if turtle.getItemDetail(i).name:find(thrash, 1 ,true) then
+                                turtle.select(i)
+                                turtle.drop()
+                                thrash_exists = true
+                            end
                         end
                     end
                 end
